minimal_requirements.txt
flask==3.0.0
flask-cors==4.0.0
neo4j==5.15.0
python-dotenv==1.0.0
galileo>=1.16.0